refactor(git): GitDiffPanel 外移 DiffTreeSection 组件
- 3279→2724 行；GitDiffPanelTreeSection.tsx 555 行，21 个 props 原样搬迁
- isEditableTarget/isFileMutationDisabled 升 export 共享（原文件再导出面保持）
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；GitDiffPanel 11 文件 126 passed，3 失败与基线逐一同名（既有）
